refactor(components): [calendar]
* Extract calendar common logic to dedicated file.
* Reorganize code for better readability.
