+++ packages/components/calendar/src/instance.ts
@@ -1,4 +1,6 @@
 import type DateTable from './date-table.vue'
+import type Calendar from './calendar.vue'
 
 export type DateTableInstance = InstanceType<typeof DateTable>
 export type CalendarDateTableInstance = DateTableInstance
+export type CalendarInstance = InstanceType<typeof Calendar>
